# cmd/go: scale go list with GOMAXPROCS