@goto1134 2023-09-21
He wrote the collective traces for the fix:
[trace-maxproc20-fix.txt](https://github.com/golang/go/files/12681854/trace-maxproc20-fix.txt)