@goto1134 2024-08-24
The benchmark was extracted to a separate PR #69052